Fixed test crashes
