+++ com.adsbynimbus.nimbus/Runtime/Plugins/RTB/Request/Imp.cs
@@ -38,4 +38,21 @@ namespace OpenRTB.Request {
         [JsonProperty("facebook_test_ad_type", DefaultValueHandling = DefaultValueHandling.Ignore)]
         public string MetaTestAdType { get; set; }
     }
+    
+    public class ApsResponse {
+        [JsonProperty("amzn_b", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public object AmznB;
+        [JsonProperty("amzn_vid", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public object AmznVid;
+        [JsonProperty("amzn_h", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public object AmznH;
+        [JsonProperty("amznp", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public object Amznp;
+        [JsonProperty("amznrdr", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public object Amznrdr;
+        [JsonProperty("amznslots", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public object Amznslots;
+        [JsonProperty("dc", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public object Dc;
+    }
 }
